final_project/agentic_draft/requirements.txt
flask==3.0.0
openai==1.6.1
pandas==2.1.4
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.1.0

